Refactor runDecisions function to improve clarity and reusability.
